+++ crates/oxide/tests/scanner.rs
@@ -1,6 +1,5 @@
 #[cfg(test)]
 mod scanner {
-    use scanner::detect_sources::DetectSources;
     use std::process::Command;
     use std::{fs, path};
 
@@ -35,18 +34,20 @@ mod scanner {
         let base = format!("{}", dir.display());
 
         // Resolve all content paths for the (temporary) current working directory
-        let mut scanner = Scanner::new(
-            Some(DetectSources::new(base.clone().into())),
-            Some(
-                globs
-                    .iter()
-                    .map(|x| GlobEntry {
-                        base: base.clone(),
-                        pattern: x.to_string(),
-                    })
-                    .collect(),
-            ),
-        );
+        let mut sources: Vec<GlobEntry> = globs
+            .iter()
+            .map(|x| GlobEntry {
+                base: base.clone(),
+                pattern: x.to_string(),
+            })
+            .collect();
+
+        sources.push(GlobEntry {
+            base: base.clone(),
+            pattern: "**/*".to_string(),
+        });
+
+        let mut scanner = Scanner::new(Some(sources));
 
         let candidates = scanner.scan();
 
@@ -65,17 +66,20 @@ mod scanner {
             ));
         }
 
+        let parent_dir = format!(
+            "{}{}",
+            fs::canonicalize(&base).unwrap().display(),
+            path::MAIN_SEPARATOR
+        );
+
         paths = paths
             .into_iter()
             .map(|x| {
-                let parent_dir = format!("{}{}", &base.to_string(), path::MAIN_SEPARATOR);
-                x.replace(&parent_dir, "")
-                    // Normalize paths to use unix style separators
-                    .replace('\\', "/")
+                x.replace(&parent_dir, "").replace('\\', "/") // Normalize paths to use unix style separators
             })
             .collect();
 
-        // Sort the output for easier comparison (depending on internal datastructure the order
+        // Sort the output for easier comparison (depending on internal data structure the order
         // _could_ be random)
         paths.sort();
 
@@ -98,7 +102,7 @@ mod scanner {
             ("b.html", None),
             ("c.html", None),
         ]);
-        assert_eq!(globs, vec!["a.html", "b.html", "c.html", "index.html"]);
+        assert_eq!(globs, vec!["*", "a.html", "b.html", "c.html", "index.html"]);
     }
 
     #[test]
@@ -110,7 +114,7 @@ mod scanner {
             ("b.html", None),
             ("c.html", None),
         ]);
-        assert_eq!(globs, vec!["a.html", "c.html", "index.html"]);
+        assert_eq!(globs, vec!["*", "a.html", "c.html", "index.html"]);
     }
 
     #[test]
@@ -124,6 +128,7 @@ mod scanner {
         assert_eq!(
             globs,
             vec![
+                "*",
                 "index.html",
                 "public/a.html",
                 "public/b.html",
@@ -148,6 +153,7 @@ mod scanner {
         assert_eq!(
             globs,
             vec![
+                "*",
                 "index.html",
                 "public/a.html",
                 "public/b.html",
@@ -170,7 +176,7 @@ mod scanner {
             ("public/b.html", None),
             ("public/c.html", None),
         ]);
-        assert_eq!(globs, vec!["index.html", "public/c.html",]);
+        assert_eq!(globs, vec!["*", "index.html", "public/c.html",]);
     }
 
     #[test]
@@ -181,7 +187,7 @@ mod scanner {
             ("src/b.html", None),
             ("src/c.html", None),
         ]);
-        assert_eq!(globs, vec![
+        assert_eq!(globs, vec!["*",
             "index.html",
             "src/**/*.{aspx,astro,cjs,cts,eex,erb,gjs,gts,haml,handlebars,hbs,heex,html,jade,js,jsx,liquid,md,mdx,mjs,mts,mustache,njk,nunjucks,php,pug,py,razor,rb,rhtml,rs,slim,svelte,tpl,ts,tsx,twig,vue}",
             "src/a.html",
@@ -198,7 +204,7 @@ mod scanner {
             ("b.png", None),
             ("c.lock", None),
         ]);
-        assert_eq!(globs, vec!["index.html"]);
+        assert_eq!(globs, vec!["*", "index.html"]);
     }
 
     #[test]
@@ -209,7 +215,7 @@ mod scanner {
             ("b.sass", None),
             ("c.less", None),
         ]);
-        assert_eq!(globs, vec!["index.html"]);
+        assert_eq!(globs, vec!["*", "index.html"]);
     }
 
     #[test]
@@ -219,7 +225,7 @@ mod scanner {
             ("package-lock.json", None),
             ("yarn.lock", None),
         ]);
-        assert_eq!(globs, vec!["index.html"]);
+        assert_eq!(globs, vec!["*", "index.html"]);
     }
 
     #[test]
@@ -270,6 +276,7 @@ mod scanner {
         assert_eq!(
             globs,
             vec![
+                "*",
                 "bar.html",
                 "baz.html",
                 "foo.html",
@@ -358,7 +365,7 @@ mod scanner {
                 // detection.
                 ("foo.styl", Some("content-['foo.styl']")),
             ],
-            vec!["*.styl"],
+            vec!["foo.styl"],
         )
         .1;