+++ packages/@tailwindcss-postcss/src/index.ts
@@ -134,11 +134,23 @@ function tailwindcss(opts: PluginOptions = {}): AcceptedPlugin {
           }
 
           if (context.scanner === null || rebuildStrategy === 'full') {
+            let sources = (() => {
+              // Disable auto source detection
+              if (context.compiler.root === 'none') {
+                return []
+              }
+
+              // No root specified, use the base directory
+              if (context.compiler.root === null) {
+                return [{ base, pattern: '**/*' }]
+              }
+
+              // Use the specified root
+              return [context.compiler.root]
+            })().concat(context.compiler.globs)
+
             // Look for candidates used to generate the CSS
-            context.scanner = new Scanner({
-              detectSources: { base },
-              sources: context.compiler.globs,
-            })
+            context.scanner = new Scanner({ sources })
           }
 
           env.DEBUG && console.time('[@tailwindcss/postcss] Scan for candidates')
@@ -159,13 +171,22 @@ function tailwindcss(opts: PluginOptions = {}): AcceptedPlugin {
           // giving tools like Vite or Parcel a glob that can be used to limit
           // the files that cause a rebuild to only those that match it.
           for (let { base, pattern } of context.scanner.globs) {
-            result.messages.push({
-              type: 'dir-dependency',
-              plugin: '@tailwindcss/postcss',
-              dir: base,
-              glob: pattern,
-              parent: result.opts.from,
-            })
+            if (pattern === '') {
+              result.messages.push({
+                type: 'dependency',
+                plugin: '@tailwindcss/postcss',
+                file: base,
+                parent: result.opts.from,
+              })
+            } else {
+              result.messages.push({
+                type: 'dir-dependency',
+                plugin: '@tailwindcss/postcss',
+                dir: base,
+                glob: pattern,
+                parent: result.opts.from,
+              })
+            }
           }
 
           env.DEBUG && console.time('[@tailwindcss/postcss] Build CSS')